Create Refugee Council of Australia.js
This is a web scraper for the Refugee Council of Australia, a not-for-profit site that includes a lot of information and publications. It's a website I have editing privileges for (as a former staff member) but wanted to add some extra metadata that I wasn't able to control. It's based on the Embedded Metadata translator with some edits to metadata, including changes to report types, tags, and removing some of the unnecessary metadata and adding multi-item scraping.