Create Australian Bureau of Statistics.js
This is a web scraper for single items of the Australian Bureau of Statistics site. It is based on the Embedded Metadata translator and replaces messy metadata with cleaned-up versions. The website doesn't distinguish well between multiple item pages and others at present, so this is limited to scraping only individual statistical items (the website is in the process of being refreshed so the new version is likely to be more compliant).