Australian Institute of Health and Welfare
This is a web scraper for the Australian Institute of Health and Welfare, an Australian research body. It scrapes only reports (as data sources do not have separate webpages but are attachment files). Based on the Embedded Metadata translator, it changes the item type to reports, adds the date, and fixes the title and publisher.